// src/app/Models/BusinessUnit.ts
export class BusinessUnit {
  businessUnitName: string;
  createdBy: string;
  createdByName: string;
  createdDateTime: string;
  id: string;
  updatedBy: string;
  updatedByName: string;
  updatedDateTime: string;
  interviewPanelList: any;
  keyword: string;
  buShortName: string;
  fromDateTime: string;
  toDateTime: string;
  managerId: string;
  managerName:string;
}