// src/app/Models/Customers.ts
export class customer {
  businessUnitId: string;
  businessUnitName: string;
  createdBy: string;
  createdByName: string;
  createdDateTime: string;
  customerName: string;
  id: string;
  location: string;
  region: string;
  updatedBy: string;
  updatedByName: string;
  updatedDateTime: string;
  keyword: string;
  customerShortName: string;
  recRoleDtoList: string[];
  recRoleIds: string[];
  status: string;
  fromDateTime:string
  toDateTime:string
}