// src/app/Models/jobReqSearch.ts
export class jobReqSearch {
  buId: string;
  buName: string;
  closedJobRequest: string;
  closedOpenings: string;
  contractDuration: string;
  createdBy: string;
  createdByName: string;
  createdDateTime: string;
  currency: string;
  customerId: string;
  customerName: string;
  customerNameInDetail: string;
  daysLeft: string;
  daysSpent: string;
  daysTotal: string;
  employmentType: string;
  holdJobRequest: string;
  id: string;
  jobDescription: string;
  jobReqStatus: string;
  location: string;
  mandatorySkills: string;
  message: string;
  monthOrYear: string;
  noOfOpenings: string;
  openJobRequest: string;
  optionalSkills: string;
  payFrequency: string;
  payRange: string;
  placementFor: string;
  progress: string;
  progressJobRequest: string;
  projectStartDate: string;
  recruiterId: string;
  recruiterName: string;
  referenceNumber: string;
  remoteOption: string;
  requesterId: string;
  requesterName: string;
  roleId: string;
  roleName: string;
  runningNumber: string;
  status: string;
  totalJobRequest: string;
  updatedBy: string;
  updatedByName: string;
  updatedDateTime: string;
  vendorId: string;
  year: string;
  toDateTime: string;
  fromDateTime: string;
  employeeId: string;
  vendorPriority: string;
  jrNumber: string;
}